+++ internal/configuration/curves.go
@@ -12,10 +12,10 @@ const (
 )
 
 type LinearCurveConfig struct {
-	Sensor  string      `json:"sensor"`
-	MinTemp int         `json:"minTemp"`
-	MaxTemp int         `json:"maxTemp"`
-	Steps   map[int]int `json:"steps"`
+	Sensor string      `json:"sensor"`
+	Min    int         `json:"min"`
+	Max    int         `json:"max"`
+	Steps  map[int]int `json:"steps"`
 }
 
 const (
