+++ internal/configuration/sensors.go
@@ -1,0 +1,7 @@
+package configuration
+
+type SensorConfig struct {
+	Id       string `json:"id"`
+	Platform string `json:"platform"`
+	Index    int    `json:"index"`
+}
